# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ArenaReplay API",
    "description": "Instant replay clips for beach courts. All times are UTC (ISO 8601).",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Health",
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/status": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "System Status",
        "description": "Live state of every service and court camera, from the service heartbeats.",
        "operationId": "system_status_api_v1_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/courts": {
      "get": {
        "tags": [
          "courts"
        ],
        "summary": "List Courts",
        "operationId": "list_courts_api_v1_courts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CourtOut"
                  },
                  "type": "array",
                  "title": "Response List Courts Api V1 Courts Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clips": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "List Clips",
        "description": "Ready clips, newest first, with cursor pagination.",
        "operationId": "list_clips_api_v1_clips_get",
        "parameters": [
          {
            "name": "court_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Court Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Local calendar day (YYYY-MM-DD) in the server timezone",
              "title": "Date"
            },
            "description": "Local calendar day (YYYY-MM-DD) in the server timezone"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "next_cursor from the previous page",
              "title": "Cursor"
            },
            "description": "next_cursor from the previous page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clips/{clip_id}": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Get Clip",
        "description": "Any status: lets a client poll a clip until it is ready.",
        "operationId": "get_clip_api_v1_clips__clip_id__get",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clips/{clip_id}/video": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Clip Video",
        "operationId": "clip_video_api_v1_clips__clip_id__video_get",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "video/mp4": {}
            }
          },
          "206": {
            "description": "Partial content"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clips/{clip_id}/thumbnail": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Clip Thumbnail",
        "operationId": "clip_thumbnail_api_v1_clips__clip_id__thumbnail_get",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/jpeg": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clips/{clip_id}/download": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Clip Download",
        "operationId": "clip_download_api_v1_clips__clip_id__download_get",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "video/mp4": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClipOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "court_id": {
            "type": "string",
            "title": "Court Id",
            "examples": [
              "court1"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ClipStatus"
          },
          "triggered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Triggered At",
            "description": "When the trigger fired (UTC)"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At",
            "description": "Real time of the first frame (UTC)"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At",
            "description": "Real time of the last frame (UTC)"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S",
            "examples": [
              33.0
            ]
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url",
            "description": "Streamable MP4 (supports HTTP Range); null until ready"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url",
            "description": "JPEG thumbnail, if any"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url",
            "description": "Same MP4 with Content-Disposition: attachment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "court_id",
          "status",
          "triggered_at",
          "start_at",
          "end_at"
        ],
        "title": "ClipOut"
      },
      "ClipPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClipOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Pass as ?cursor= to get the next (older) page; null at the end"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ClipPage"
      },
      "ClipStatus": {
        "type": "string",
        "enum": [
          "processing",
          "ready",
          "failed"
        ],
        "title": "ClipStatus"
      },
      "ClipperStatusOut": {
        "properties": {
          "trigger": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger",
            "examples": [
              "keyboard"
            ]
          },
          "queue_size": {
            "type": "integer",
            "title": "Queue Size",
            "default": 0
          },
          "processing": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing"
          },
          "processing_court": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Court"
          },
          "ready_count": {
            "type": "integer",
            "title": "Ready Count",
            "description": "Clips ready since the clipper started",
            "default": 0
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count",
            "description": "Clips failed since the clipper started",
            "default": 0
          },
          "last_ready_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Ready At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_error_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error At"
          }
        },
        "type": "object",
        "title": "ClipperStatusOut"
      },
      "CourtOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "court1"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Court 1"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CourtOut"
      },
      "CourtStatusOut": {
        "properties": {
          "court_id": {
            "type": "string",
            "title": "Court Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "capture": {
            "type": "string",
            "enum": [
              "connecting",
              "recording",
              "reconnecting",
              "starting",
              "stopped",
              "down",
              "unknown"
            ],
            "title": "Capture"
          },
          "since": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since",
            "description": "When `capture` last changed"
          },
          "last_segment_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Segment At"
          },
          "segment_age_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment Age S",
            "description": "Seconds since the newest segment started"
          },
          "restarts": {
            "type": "integer",
            "title": "Restarts",
            "default": 0
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_error_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error At"
          }
        },
        "type": "object",
        "required": [
          "court_id",
          "name",
          "capture"
        ],
        "title": "CourtStatusOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthOut": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "examples": [
              "ok"
            ]
          },
          "database": {
            "type": "boolean",
            "title": "Database"
          }
        },
        "type": "object",
        "required": [
          "status",
          "database"
        ],
        "title": "HealthOut"
      },
      "ServiceStatusOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "capture"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "running",
              "stopped",
              "down",
              "unknown"
            ],
            "title": "State",
            "description": "running; stopped (clean shutdown); down (heartbeat lost: crash or killed); unknown (never ran here)"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Last heartbeat"
          }
        },
        "type": "object",
        "required": [
          "name",
          "state"
        ],
        "title": "ServiceStatusOut"
      },
      "SystemStatusOut": {
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "ok",
              "warning",
              "error"
            ],
            "title": "Level",
            "description": "error: a service or the database is down; warning: a court is not recording or a clip failed recently; ok otherwise"
          },
          "problems": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Problems",
            "description": "Human-readable reasons for the level"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          },
          "database": {
            "type": "boolean",
            "title": "Database"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/ServiceStatusOut"
            },
            "type": "array",
            "title": "Services"
          },
          "courts": {
            "items": {
              "$ref": "#/components/schemas/CourtStatusOut"
            },
            "type": "array",
            "title": "Courts"
          },
          "clipper": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClipperStatusOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when the clipper is not running"
          }
        },
        "type": "object",
        "required": [
          "level",
          "problems",
          "checked_at",
          "database",
          "services",
          "courts"
        ],
        "title": "SystemStatusOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}